fixing brew warning in goreleaser, auto-generating release properties, move build to GHA
Pulling some of the logic from the Makefile into the CLI here.

No special support for GitHub actions yet, but this shouldn't break any situation in which folks are using the Makefile and its special GHA logic to release.

```

$ go run ./cli release create --auto
  • Reading Environment ✓

Prepared to create release with defaults:

    yaml-dir        "./manifests"
    promote         "auto-release"
    version         "auto-release-180c2e-dirty"
    release-notes   "CLI release by dex on 11 Aug 20 15:44 CDT"
    ensure-channel  true

Create release with these properties? [Y/n] █

  • Reading manifests from ./manifests ✓
  • Creating Release ✓
    • SEQUENCE: 128
  • Promoting ✓
    • Channel 1fy7ktlXka3WylbLIR3w218CO1G successfully set to release 128

```