fix(cmx): return proper error on vm deadline and print vm info (#551)
`createVM` was expecting a `ErrWaitDurationExceeded`, where `waitForVM`
was returning an ErrVMWaitDurationExceeded. Additionally, only info
about running VMs was returned when the deadline was exceeded. This
behaves differently than `waitForCluster` which prints out info about
clusters that did not enter the running state.

This changes waitForVM to return the proper error when the deadline is
exceeded as well as returning info about all VMs when that occurs. This
helps provide better feedback to the user when their VMs fail to come up
in the allotted time.